Add import key pair to import cluster
